+++ mathics/builtin/directories/__init__.py
@@ -1,11 +1,10 @@
 """
-Input/Output, Files, and Filesystem
-
+Directories and Directory Operations
 """
 # Note: everything in this module is not loaded if environment
 # variable ENABLE_FILES_MODULE is False.  Here we do not want to
-# include any built-in commands that can write to the filesystem.
+# include any built-in commands that can write to the filesytesm.
 
 
 # This tells documentation how to sort this module
-sort_order = "mathics.builtin.input-output-files-and-filesystem"
+sort_order = "mathics.builtin.directories-and-directory-operations"
