+++ mathics/eval/directories.py
@@ -1,0 +1,3 @@
+import os
+
+SYS_ROOT_DIR = "/" if os.name == "posix" else "\\"
